+++ opentelemetry-appender-tracing/src/lib.rs
@@ -30,9 +30,9 @@
 //! tracing = ">=0.1.40"
 //! tracing-core = { version = ">=0.1.33" }
 //! tracing-subscriber = { version = "0.3", features = ["registry", "std", "fmt"] }
-//! opentelemetry = { version = "0.28", features = ["logs"] }
-//! opentelemetry-sdk = { version = "0.28", features = ["logs"] }
-//! opentelemetry-appender-tracing = { version = "0.28.1" }
+//! opentelemetry = { version = "0.31", features = ["logs"] }
+//! opentelemetry-sdk = { version = "0.31", features = ["logs"] }
+//! opentelemetry-appender-tracing = { version = "0.31.1" }
 //! ```
 //!
 //! ### 2. Set Up the OpenTelemetry Logger Provider